+++ src/structured-data/lib.js
@@ -82,9 +82,9 @@ export async function getIssuesFromGSC(finalUrl, context, pages) {
         return;
       }
 
-      richResults.detectedItems.forEach((type) => {
-        type.items.forEach((item) => {
-          item.issues.forEach((issue) => {
+      richResults.detectedItems?.forEach((type) => {
+        type?.items?.forEach((item) => {
+          item?.issues?.forEach((issue) => {
             const rootType = entityMapping[type.richResultType];
             if (!rootType) {
               log.warn(`SDA: Skipping GSC issue, because cannot map GSC type "${type.richResultType}" to schema.org type.`);
@@ -187,15 +187,19 @@ export async function getIssuesFromScraper(context, pages, scrapeCache) {
 
     const schemaOrgPath = join(
       process.cwd(),
-      'src',
-      'structured-data',
+      'static',
       'schemaorg-current-https.jsonld',
     );
 
     const validator = new StructuredDataValidator(schemaOrgPath);
-    const validatorIssues = (await validator.validate(waeResult))
-      // For now, ignore issues with severity lower than ERROR
-      .filter((issue) => issue.severity === 'ERROR');
+    let validatorIssues = [];
+    try {
+      validatorIssues = (await validator.validate(waeResult))
+        // For now, ignore issues with severity lower than ERROR
+        .filter((issue) => issue.severity === 'ERROR');
+    } catch (e) {
+      log.error(`SDA: Failed to validate structured data for ${page}.`, e);
+    }
     for (const issue of validatorIssues) {
       // Only add if same issue for the same source does not exist already.
       // This can happen e.g. if a field is missing for every item in a list.