+++ test/audits/structured-data/structured-data.test.js
@@ -75,7 +75,7 @@ describe('Structured Data Audit', () => {
       isHandlerEnabledForSite: sinon.stub().returns(true),
     };
     s3ClientStub = {
-      send: sinon.stub(),
+      send: sinon.mock(),
       getObject: sinon.stub(),
     };
     firefallClientStub = {
@@ -156,6 +156,26 @@ describe('Structured Data Audit', () => {
           success: false,
         },
       });
+    });
+
+    it('filters out files from top pages', async () => {
+      context.dataAccess.SiteTopPage.allBySiteIdAndSourceAndGeo = sinon.stub().resolves([
+        createPageStub('https://example.com/product/1'),
+        createPageStub('https://example.com/product/2.pdf'),
+        createPageStub('https://example.com/product/8.xlsx'),
+      ]);
+
+      context.dataAccess.Opportunity.allBySiteIdAndStatus
+        .resolves([context.dataAccess.Opportunity]);
+      context.dataAccess.Opportunity.getSuggestions.resolves([]);
+      context.dataAccess.Opportunity.getId.returns('opportunity-id');
+      context.dataAccess.Opportunity.getType.returns('structured-data');
+      context.dataAccess.Opportunity.addSuggestions.resolves(structuredDataSuggestions);
+
+      await runAuditAndGenerateSuggestions(context);
+      expect(s3ClientStub.send.calledOnce).to.equal(true);
+      const scrapeRequests = s3ClientStub.send.getCalls().map((call) => call.args[0].input.Key);
+      expect(scrapeRequests).to.deep.equal(['scrapes/123/product/1/scrape.json']);
     });
 
     it('runs a full audit', async () => {
@@ -439,7 +459,7 @@ describe('Structured Data Audit', () => {
       s3ClientStub.send.rejects(new Error('Failed to fetch S3 object'));
 
       await generateSuggestionsData(finalUrl, auditData, context, scrapeCache);
-      expect(context.log.error).to.have.been.calledWith('SDA: Could not find scrape for /product/1. Make sure that scrape-top-pages did run.');
+      expect(context.log.error).to.have.been.calledWith('SDA: Could not find scrape for https://example.com/product/1 at /product/1. Make sure that scrape-top-pages did run.');
       expect(firefallClientStub.fetchChatCompletion).to.not.have.been.called;
     });